feat: adds start mc to macos

+++ init/mac.sh
@@ -1,24 +1,52 @@
 #!/bin/bash
 
 install_packages() {
-	echo -e "\nInstalling Docker for MacOS..."
+	read -r -p "\nInstall Docker & docker-compose? [y/N] " answer
+	if [[ "$answer" != y ]] && [[ "$answer" != Y ]]; then
+		echo -e "\nSkipping Docker systemctl setup...\n"
+	else
+		echo -e "\nInstalling Docker for MacOS..."
 
-	declare -a packages=(
-		"docker"
-		"docker-compose"
-	)
+		declare -a packages=(
+			"docker"
+			"docker-compose"
+		)
 
-	brew install ${packages[@]}
-	brew cask install docker
+		brew install ${packages[@]}
+		brew cask install docker
 
-	echo -e "DONE.\n"
+		echo -e "DONE.\n"
+	fi
 }
 
 setup_homebrew() {
-	if [[ ! $(which brew) ]] && [[ ! $(brew --version &> /dev/null) ]]; then
-		echo -e "\nCould not find existing HOMEBREW installation."
-		echo -e "\nNow installing..."
-		/usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+	read -r -p "\nInstall Homebrew? [y/N] " answer
+	if [[ "$answer" != y ]] && [[ "$answer" != Y ]]; then
+		echo -e "\nSkipping Docker systemctl setup...\n"
+	else
+		if [[ ! $(which brew) ]] && [[ ! $(brew --version &> /dev/null) ]]; then
+			echo -e "\nCould not find existing HOMEBREW installation."
+			echo -e "\nNow installing..."
+			/usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+		fi
+	fi
+}
+
+
+run_mc() {
+	read -r -p "Start Mission Control container? [y/N] " answer
+	if [[ "$answer" != y ]] && [[ "$answer" != Y ]]; then
+		echo -e "\nSkipping docker-compose up --build -d..."
+	else
+		echo -e "\nRunning Mission Control backend"
+		cd ../ && \
+		echo -e "\nSourcing .env"
+		source sourceme.sh && \
+		echo -e "\nStarting container..."
+		docker-compose up --build -d
+		echo
+		echo -e "\nRun 'prisma deploy' & 'prisma seed' from"
+		echo -e "\nthe root directory to get started"
 	fi
 }
 
@@ -27,8 +55,7 @@ setup_macos() {
 
 	setup_homebrew
 	install_packages
-
-	echo -e "DONE.\n"
+	run_mc
 }
 
 setup_macos
